chore: move members of magnifier feature into a seperate files when possible (not final yet)
